requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.0
cairocffi==1.5.1
CairoSVG==2.7.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.2.1
conda==22.9.0
conda-package-handling @ file:///opt/conda/conda-bld/conda-package-handling_1663598473529/work
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1665612644927/work
cssselect2==0.7.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@1523b3e9929a06d798871eb9afc4c9f770743baf
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
fvcore==0.1.5.post20221221
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-cloud-vision==3.3.1
googleapis-common-protos==1.58.0
grad-cam==1.4.6
grpcio==1.51.1
grpcio-status==1.51.1
hydra-core==1.3.1
idna @ file:///croot/idna_1666125576474/work
imageio==2.25.0
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.0.0
iopath==0.1.9
ipykernel==6.16.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==7.3.4
jupyter-server==1.23.5
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.5
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.6.3
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.14.0
pynvml==11.4.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
reportlab==3.6.12
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
svglib==1.5.1
svgpathtools==1.5.1
svgwrite==1.4.3
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.2.3
tinycss2==1.2.1
tomli==2.0.1
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
ttach==0.0.3
typing_extensions==4.2.0
urllib3==1.26.9
Wand==0.6.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.1.2
widgetsnbextension==4.0.5
yacs==0.1.8
zipp==3.8.0
